rag/requirements.txt
# RAG System Dependencies for MBTI Documentation

# Core RAG Framework
langchain==0.1.0
langchain-community==0.0.10

# Vector Database
chromadb==0.4.22

# Embeddings
sentence-transformers==2.2.2
openai==1.7.2

# Document Processing
pypdf==3.17.4
python-dotenv==1.0.0
tiktoken==0.5.2

# Web Interface
streamlit==1.29.0
streamlit-chat==0.1.1

# Utilities
numpy==1.24.3
pandas==2.1.4
tqdm==4.66.1

# Optional: Advanced Features
# faiss-cpu==1.7.4  # For faster similarity search
# cohere==4.37  # For Cohere embeddings
